Создал файлы Course.java и Team.java (пустые, с заданиями) Подправил имена пакеджей (с маленькой буквы, так вроде бы правильней)
