--- JavaCore/Lesson1(OOP)/src/Lesson_1/Marathon/Main.java
@@ -1,17 +1,0 @@
-package Lesson_1.Marathon;
-
-public class Main {
-    public static void main(String[] args) {
-        Competitor[] competitors = {new Human("Боб"), new Cat("Барсик"), new Dog("Бобик")};
-        Obstacle[] course = {new Cross(80), new Water(2), new Wall(1), new Cross(120)};
-        for (Competitor c : competitors) {
-            for (Obstacle o : course) {
-                o.doIt(c);
-                if (!c.isOnDistance()) break;
-            }
-        }
-        for (Competitor c : competitors) {
-            c.info();
-        }
-    }
-}
