+++ JavaCore/Lesson1(OOP)/src/lesson_1/marathon/Cource.java
@@ -1,0 +1,10 @@
+package lesson_1.marathon;
+
+/**
+ *3. Добавить класс Course (полоса препятствий),
+ * в котором будут находиться массив препятствий и метод,
+ * который будет просить команду пройти всю полосу;
+ */
+
+public class Cource {
+}
