+++ JavaCore/Lesson1(OOP)/src/lesson_1/marathon/Team.java
@@ -1,0 +1,11 @@
+package lesson_1.marathon;
+
+/**
+ * 2. Добавить класс Team, который будет содержать название команды,
+ * массив из четырех участников (в конструкторе можно сразу указыватьвсех участников ),
+ * метод для вывода информации о членах команды, прошедших дистанцию,
+ * метод вывода информации обо всех членах команды.
+ */
+
+public class Team {
+}
