+++ JavaCore/Lesson1(OOP)/src/lesson_1/marathon/competitors/Human.java
@@ -1,4 +1,4 @@
-package Lesson_1.Marathon;
+package lesson_1.marathon.competitors;
 
 public class Human implements Competitor {
     String name;
