+++ JavaCore/Lesson1(OOP)/src/lesson_1/marathon/obstacles/Cross.java
@@ -1,4 +1,6 @@
-package Lesson_1.Marathon;
+package lesson_1.marathon.obstacles;
+
+import lesson_1.marathon.competitors.Competitor;
 
 public class Cross extends Obstacle {
     int length;
